feat: Allow registering players as free agents
Update the player registration form in the admin view to allow users to register players without assigning them to a club. This is achieved by changing the "Club" select input to be optional and adding a "No Club (Free Agent)" option. A new informational message is displayed when no club is selected.