feat: Add Man of the Match field to fixtures
Introduces a 'manOfTheMatchId' field to the Fixture type to store the ID of the player awarded Man of the Match. Updates the Fixture form in the admin view to allow selection of the Man of the Match from eligible players and displays this information in the Score Sheet Modal. The Player Detail View now also shows a list of fixtures where the player was awarded Man of the Match.